fix: 이벤트 로그 조회 시 이벤트 타입 수정 - EventPortIntegrationTest에서 ORDER_CREATED를 ORDER_COMPLETED로 변경하여 이벤트 로그 조회를 정확하게 수정 - KafkaCouponEventIntegrationTest에서 쿠폰의 totalQuantity 타입을 Integer에서 Long으로 변경하고, updateIssuedQuantity 메서드 호출 시 int로 변환하여 NPE 방지

+++ src/test/java/kr/hhplus/be/server/integration/event/KafkaCouponEventIntegrationTest.java
@@ -218,9 +218,9 @@ class KafkaCouponEventIntegrationTest extends IntegrationTestBase {
         txTemplate.execute(status -> {
             Coupon coupon = couponRepository.findById(testCoupon.getId()).orElseThrow();
             // NPE 방지: totalQuantity가 null일 수 있으므로 안전하게 처리
-            Integer totalQty = coupon.getTotalQuantity();
+            Long totalQty = coupon.getTotalQuantity();
             if (totalQty != null) {
-                coupon.updateIssuedQuantity(totalQty);
+                coupon.updateIssuedQuantity(totalQty.intValue());
             } else {
                 // totalQuantity가 null이면 0으로 간주하고 재고를 0으로 설정
                 coupon.updateIssuedQuantity(0);
